Rename files from heap to bin_heap
The use of 'heap' for labelling the data structure clashes with dynamic
memory 'heap'. This is purely a naming convention change, and should have
no affect on implementation.
