+++ src/native.rs
@@ -2,14 +2,14 @@ mod arc_list;
 mod atomic_waker;
 mod delay;
 mod global;
-mod heap;
-mod heap_timer;
+mod bin_heap;
+mod bin_heap_timer;
 mod timer;
 
 use self::arc_list::{ArcList, Node};
 use self::atomic_waker::AtomicWaker;
-use self::heap::{Heap, Slot};
-use self::heap_timer::HeapTimer;
+use self::bin_heap::{Heap, Slot};
+use self::bin_heap_timer::HeapTimer;
 use self::timer::{ScheduledTimer, Timer, TimerHandle};
 
 pub use self::delay::Delay;
